refactor: move route to class
